+++ packages/gamut-kit/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.6.543](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.542...@codecademy/gamut-kit@0.6.543) (2025-09-15)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+### [0.6.542](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.541...@codecademy/gamut-kit@0.6.542) (2025-09-12)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
 ### [0.6.541](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.540...@codecademy/gamut-kit@0.6.541) (2025-09-09)
 
 **Note:** Version bump only for package @codecademy/gamut-kit